+++ InBrainSdk/Assets/InBrain/Scripts/API/Entities/InBrainSurvey.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace InBrain
@@ -7,42 +8,54 @@ namespace InBrain
 	public class InBrainSurvey
 	{
 		[SerializeField] public string id;
+		[SerializeField] public string searchId;
 		[SerializeField] public long rank;
 		[SerializeField] public long time;
 		[SerializeField] public float value;
 		[SerializeField] public bool currencySale;
 		[SerializeField] public float multiplier;
+		[SerializeField] public List<InBrainSurveyCategory> categories;
+		[SerializeField] public InBrainSurveyConversionLevel conversionLevel;
 
-		public InBrainSurvey(string id, long rank, long time, float value, bool currencySale, float multiplier)
+		public InBrainSurvey(string id, string searchId, long rank, long time, float value, bool currencySale, float multiplier,
+			List<InBrainSurveyCategory> categories, InBrainSurveyConversionLevel conversionLevel)
 		{
 			this.id = id;
+			this.searchId = searchId;
 			this.rank = rank;
 			this.time = time;
 			this.value = value;
 			this.currencySale = currencySale;
 			this.multiplier = multiplier;
+			this.categories = categories;
+			this.conversionLevel = conversionLevel;
 		}
 
 		public static InBrainSurvey FromAJO(AndroidJavaObject ajo)
 		{
 			return new InBrainSurvey(ajo.Get<string>("id"),
+				ajo.Get<string>("searchId"),
 				ajo.Get<long>("rank"),
 				ajo.Get<long>("time"),
 				ajo.Get<float>("value"),
 				ajo.Get<bool>("currencySale"),
-				ajo.Get<float>("multiplier"));
+				ajo.Get<float>("multiplier"),
+				ajo.GetAJO("categories").FromJavaList(category => category.FromSurveyCategoryAJO()),
+				ajo.GetAJO("conversionLevel").FromSurveyConversionLevelAJO());
 		}
 
 		public AndroidJavaObject ToAJO()
 		{
 			return Application.platform == RuntimePlatform.Android
-				? new AndroidJavaObject("com.inbrain.sdk.model.Survey", id, rank, time, value, currencySale, multiplier)
+				? new AndroidJavaObject("com.inbrain.sdk.model.Survey",
+					id, rank, time, value, currencySale, multiplier, conversionLevel.ToAJO(), searchId, categories.ToJavaList(category => category.ToAJO()))
 				: null;
 		}
 
 		public override string ToString()
 		{
-			return string.Format("id: {0}, rank: {1}, time: {2}, value: {3}, currencySale: {4}, multiplier: {5}", id, rank, time, value, currencySale, multiplier);
+			return string.Format("id: {0}, searchId: {1}, rank: {2}, time: {3}, value: {4}, currencySale: {5}, multiplier: {6}, conversionLevel: {7}, categories: {8}",
+				id, searchId, rank, time, value, currencySale, multiplier, conversionLevel, string.Join(";", categories.ToArray()));
 		}
 	}
 }